new logistic circuit likelihood api

+++ test/Juice/Probabilistic/CircuitQueriesTest.jl
@@ -6,8 +6,7 @@ if endswith(@__FILE__, PROGRAM_FILE)
 using Test
 using .Juice
 
-function generate_data_all(N)
-    # N = 4;
+function generate_data_all(N::Int32)
     data_all = transpose(parse.(Bool, split(bitstring(0)[end-N+1:end], "")));
     for mask = 1: (1<<N) - 1
         data_all = vcat(data_all,
